+++ ndk/src/signers/nip46/index.ts
@@ -181,6 +181,7 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
                 (response: NDKRpcResponse) => {
                     if (response.result === "ack") {
                         this.getPublicKey().then((pubkey) => {
+                            this.userPubkey = pubkey;
                             this._user = new NDKUser({ pubkey });
                             resolve(this._user);
                         });
@@ -239,8 +240,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
         value: string,
         method: "nip04" | "nip44"
     ): Promise<string> {
-        this.debug("asking for encryption");
-
         const promise = new Promise<string>((resolve, reject) => {
             if (!this.bunkerPubkey) throw new Error("Bunker pubkey not set");
 
@@ -267,8 +266,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
         value: string,
         method: "nip04" | "nip44"
     ): Promise<string> {
-        this.debug("asking for decryption");
-
         const promise = new Promise<string>((resolve, reject) => {
             if (!this.bunkerPubkey) throw new Error("Bunker pubkey not set");
 
@@ -291,8 +288,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
     }
 
     public async sign(event: NostrEvent): Promise<string> {
-        this.debug("asking for a signature");
-
         const promise = new Promise<string>((resolve, reject) => {
             if (!this.bunkerPubkey) throw new Error("Bunker pubkey not set");
 
@@ -302,7 +297,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
                 [JSON.stringify(event)],
                 24133,
                 (response: NDKRpcResponse) => {
-                    this.debug("got a response", response);
                     if (!response.error) {
                         const json = JSON.parse(response.result);
                         resolve(json.sig);
@@ -329,7 +323,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
         email?: string
     ): Promise<Hexpubkey> {
         await this.startListening();
-        this.debug("asking to create an account");
         const req: string[] = [];
 
         if (username) req.push(username);
@@ -345,7 +338,6 @@ export class NDKNip46Signer extends EventEmitter implements NDKSigner {
                 req,
                 NDKKind.NostrConnect,
                 (response: NDKRpcResponse) => {
-                    this.debug("got a response", response);
                     if (!response.error) {
                         const pubkey = response.result;
                         resolve(pubkey);