+++ ndk/src/signers/nip46/rpc.ts
@@ -11,6 +11,8 @@ import {
     type NDKFilter,
     type NDKSubscription,
 } from "../../subscription";
+import { NDKRelayAuthPolicies } from "../../relay/auth-policies";
+import { NDKPool } from "../../relay/pool";
 
 export interface NDKRpcRequest {
     id: string;
@@ -33,12 +35,31 @@ export class NDKNostrRpc extends EventEmitter {
     private relaySet: NDKRelaySet | undefined;
     private debug: debug.Debugger;
     public encryptionType: "nip04" | "nip44" = "nip04";
+    private pool: NDKPool | undefined;
 
     public constructor(ndk: NDK, signer: NDKSigner, debug: debug.Debugger, relayUrls?: string[]) {
         super();
         this.ndk = ndk;
         this.signer = signer;
-        this.relaySet = relayUrls ? NDKRelaySet.fromRelayUrls(relayUrls, ndk) : undefined;
+
+        // if we have relays, we create a separate pool for it
+        if (relayUrls) {
+            this.pool = new NDKPool(
+                relayUrls,
+                Array.from(ndk.pool.blacklistRelayUrls),
+                ndk,
+                debug.extend("rpc-pool")
+            );
+            
+            this.relaySet = new NDKRelaySet(new Set(), ndk, this.pool);
+            for (const url of relayUrls) {
+                const relay = this.pool.getRelay(url, false, false);
+                relay.authPolicy = NDKRelayAuthPolicies.signIn({ ndk, signer, debug });
+                this.relaySet.addRelay(relay);
+                relay.connect();
+            }
+        }
+
         this.debug = debug.extend("rpc");
     }
 
@@ -52,12 +73,11 @@ export class NDKNostrRpc extends EventEmitter {
                 closeOnEose: false,
                 groupable: false,
                 cacheUsage: NDKSubscriptionCacheUsage.ONLY_RELAY,
+                pool: this.pool,
             },
             this.relaySet,
             false
         );
-
-        this.debug("starting subscription");
 
         sub.on("event", async (event: NDKEvent) => {
             try {
@@ -85,19 +105,26 @@ export class NDKNostrRpc extends EventEmitter {
         // support both nip04 and nip44 encryption
         if (this.encryptionType === 'nip44' && event.content.includes('?iv=')) {
             this.encryptionType = 'nip04';
-            console.log('switching to nip04 encryption')
         } else if (this.encryptionType === 'nip04' && !event.content.includes('?iv=')) {
             this.encryptionType = 'nip44';
-            console.log('switching to nip44 encryption')
         }
 
         const remoteUser = this.ndk.getUser({ pubkey: event.pubkey });
         remoteUser.ndk = this.ndk;
-        const decryptedContent = await this.signer.decrypt(
-            remoteUser,
-            event.content,
-            this.encryptionType
-        );
+        let decryptedContent: string;
+
+        try {
+            decryptedContent = await this.signer.decrypt(
+                remoteUser,
+                event.content,
+                this.encryptionType
+            );
+        } catch (e) {
+            const otherEncryptionType = this.encryptionType === 'nip04' ? 'nip44' : 'nip04';
+            decryptedContent = await this.signer.decrypt(remoteUser, event.content, otherEncryptionType);
+            this.encryptionType = otherEncryptionType;
+        }
+        
         const parsedContent = JSON.parse(decryptedContent);
         const { id, method, params, result, error } = parsedContent;
 
@@ -128,7 +155,7 @@ export class NDKNostrRpc extends EventEmitter {
             tags: [["p", remotePubkey]],
             pubkey: localUser.pubkey,
         } as NostrEvent);
-
+        
         event.content = await this.signer.encrypt(remoteUser, event.content, this.encryptionType);
         await event.sign(this.signer);
         await event.publish(this.relaySet);