+++ ndk-mobile/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nostr-dev-kit/ndk-mobile
 
+## 0.2.1
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nostr-dev-kit/ndk@2.10.7
+    -   @nostr-dev-kit/ndk-wallet@0.3.15
+
 ## 0.2.0
 
 ### Minor Changes